fix(oriented): avoid using type aliasing
This diff refactors how we manage orientation/oriented tensors
to avoid using type aliasing, which seemed to nullify the benefits of
using distrinct types ofr distinct tensor orientations.

To this end, drop all type aliasing and always use the Tensor[]
and Field[] generic classes explicitly in the code. Also, overhaul
the creation of a Tensor[T] by having the Field[T] accept a
grap.Tensor as argument and wrapping it, rather than providing
explicit .placeholder and .constant methods in Field. This
was done out an abundance of caution and mutual distrust
between me and the Python type system.

While there rename orientation to oriented, with the motivation that
oriented.Tensor[TimeWise] reads very well and is intuitive.