fix(backend): Field[O] -= shape changing ops
This diff fixes a serious and embarrassing bug in the Field[O] type
where we allow shape-changing operations to produce tensors in the
sane field O, which is completely and utterly wrong.

For example:

```Python
m.field.total_early = m.field.reduce_sum(m.field.total_early_shift, axis=1)
```

this is completely wrong. We're using `axis=1`, so the result has a
differently dimesionality and belongs to another tensor field.