feat: add examples/assembly.py
I think this example captures the essence of the computation
performed by the dt-model, while still falling short of providing
nice abstractions to allow writing a compact model.

Because of this, and as documented into the file itself, I believe
this file showcases what could be the interface from a more abstract
model representation and its computational representation.