feat(yakof.frontend): computational-graph building (#5)
The yakof.frontend.graph package is a refactoring of the
yakof.backend.graph package focusing on just building a computation
graph.

The yakof.frontend.graph name captures more precisely what we are doing
here. When building a computation graph, we are in the "compiler"
frontend. The backend name, instead, seems more proper if reserved for
numpy operations.

Compared to yakof.backend.graph, we have distilled features and omitted
unnecessary features, such as the ability to use operations directly.

While there, also add unit tests for the yakof.frontend.graph package.