feat(yakof.numpybackend): add reference evaluator (#7)
This commit adds a tree walking reference evaluator. We strive to keep
it simple and easy to understand. We'll use more complex strategies, if
benchmarks and real use cases show it's a memory or CPU bottleneck.

While there, generally strive to improve docs.

Also, note that, to implement the numpybackend, we also need a graph
pretty printing functionality in tree.