feat(yakof.frontend): add the abstract module (#10)
This module adds abstract support for creating tensors (in the
engineering sense) and tensor spaces (ditto) using specific bases of the
possible vector spaces. We use generics to do this.

The result is that we build a category, `TensorSpace[B]`, which
implements endomorphisms, and we still need to build other abstractions
to implement the remaining morphisms between spaces.

The `Tensor[B]` is a convenience wrapper for `graph.Node` that, not only
binds `Tensor[B]` to a basis `B`, but also implements convenience infix
operations so that we could write stuff like `a + b`.