feat(frontend): add support for morphisms
A morphism is a transformation between different tensor spaces. They
transform a tensor in a given space into a tensor in another space.

Adding support for this functionality complements our existing set
of endomorphisms operations, i.e., operations that transform a tensor
a given type into a tensor of the same type.

Equipped with this functionality, we can now define bases and spaces
for domain-specific problems.

The core of this diff consists of the following concepts:

1. defining a uniform way to assign unique numbers to axes, which
is done inside morphisms.generate_canonical_axes;

2. figuring out a way to embed axes into distinct types, which
is implemented by morphisms.Basis;

3. defining classes for instantiating morphisms, currently
morphisms.ExpandDims and morphisms.ProjectUsingSum, which
map to the two common operations `np.expand_dims(axis=...)` and
`np.sum(..., axis=...)`;

4. mapping the numbers assigned to axis to relative axes
positions expected by numpy, which depend on the geometry
of the problem, which we implement with:

- `axes_expansion`

- `axes_projection`

The test suite adds exhaustive testing that those two
functions are working as intended in R^4, which should
be good enough to spot several corned cases, along
with tests ensuring results computed using the frontend
and the evaluator match direct numpy computation.