fix(TensorSpace.placeholder): require name to be present (#18)
This constructor is still relatively low level. Allowing the name to be
omitted would probably make the code more confusing, while being of
little convenience, since there is little expectation that one would
write this code directly and manually. By doing this, we resolve a TODO
left in the codebase ahead of attempting to merge this code into the
dt_model.