feat(graph): auto-assign unique ID to nodes (#21)
This diff introduces atomic integers and uses them to assign unique IDs
to graph nodes. We do this for several reasons including (1) assigning a
unique ID to help with the dashboards and (2) reordering nodes by
creation time. Specifically, soon I will commit a diff that takes
advantage of the latter.