feat(frontend): implement topological sorting (#22)
This diff introduces linearization of the leaves of the computation
graph using topological sorting and insertion order to ensure
deterministic ordering.

This feature is a fundamental building block to streamline evaluating
and debugging computation graphs. The ability to sort by insertion
order, in particular, makes it very easy to follow the operations
performed by the evaluator.