refactor(numpybackend): create module for dispatch tables
Rather than keeping the dispatch tables inside evaluator.py, move
them into a separate module, which allows sharing them, in the future,
with other implementations of the evaluator concept.

This commit serves as a hint that I am adding support for more
evaluators in the near future, since I think we need to experiment
a bit with evaluating the linearized graph.