feat(numpybackend): introduce topological executor
The executor works as long as we have topologically linearized the
graph using the yakof.frontend.linearize package.

If that is the case, the executor will evaluate a node assuming
that any dependency has already been evaluted.

As a result, the executor MUST have state, while the tree based
evaluator could or could not use state.

Since state is mandatory, we can merge the placeholders and
the state, thus further simplifying the code.