feat(yakof): add the trafficmodel package (#27)
With this diff, we add to yakof the trafficmodel package. This model is
~equivalent to the one we originally wrote as the yakof.fieldspace
package -- along with the traffic.py example -- except that this model
uses the new frontend and the numpybackend and, also, it's reusable and
tested.

This work is part of a plan to remove the yakof.fieldspace package and
to update the existing traffic.py file to use the new code.