feat(yakof): add the minisimulator package (#37)
The minisimulator package implements a naive minimal simulator allowing
us to generate the input space parameters.

The actual `dt-model` package implements a better model.

Here we need a simplified model to show that we can integrate with the
more complex code in the `dt-model` without having to deal with all the
`dt-model` complexity.