cleanup: remove the initial yakof implementation (#38)
We have rewritten all the code that matters to use the new
implementation, which is functionally equivalent for the most part and
definitely better thanks to a better understanding of what we're
actually doing with `dt-model`.

We still need to handle the minor fallout caused by this change, but
overall we're looking good.