fix(autoenum): reduce enum space size and start from zero (#44)
This diff modifies the autoenum algorithm and the related tests such
that the following holds:

1. we use 6 bits for representing an enum space, which allows
representing up to 64 distinct values;

2. we start counting from zero, which is more intuitive and allows a
more efficient usage of the available space.

These changes were developed when debugging a model where it occurred to
me that using large (e.g., `1<<20`) enum spaces makes for a rather
obscure and sad debugging experience.