feat: add dt-model as a dependency (#47)
We want to add the dt-model as a dependency, which enables us to perform
A/B testing against the current dt-model.

Note that, additionally, this dependency allows us to move selected
functionality to dt-model, such that we can incrementally hollow this
package and replace it with code written for the dt-model.