feat: add dt_model emulation layer
This diff adds a modified version of most (no pun intended) of the
dt_model that allows us to evaluate models using a codebase that is
as similar as possible to the actual dt_model.

The only aspect of dt_model that remains as a dependency is the
"scheduler" component of the architecture, i.e., the one that
ensures we're properly creating ensembles.

In general, this code is not meant to be merged upstream into the
dt_model, rather its function is to aid us in A/B testing.