feat: add dt_model emulation layer (#50)
This diff adds a modified version of most (no pun intended) of the
dt_model that allows us to evaluate models using a codebase that is as
similar as possible to the actual dt_model.

The only aspect of dt_model that remains as a dependency is the
"scheduler" component of the architecture, i.e., the one that ensures
we're properly creating ensembles.

In general, this code is not meant to be merged upstream into the
dt_model, rather its function is to aid us in A/B testing.ble

At the moment in which we're merging this code, there are some question
marks, open problems and TODOs to deal with. In any case, I feel like
the time is ripe for merging and to move forward to more aggressive A/B
testing. We'll therefore deal with the issues at a later time, using the
existing TODO comments as an atlas to know what needs fixes.

Note that, to merge this code, we also need to pull in scipy as a
dependency, since we need to create distributions.