BREAKING CHANGE(graph.py): implement infix and prefix operators
This diff implements infix and prefix operators to allow for
gradual migration of the dt_model to use typed tensors. These
operators significantly improve code readability and ergonomics
for graph construction.

We also add tests to ensure the new code is working as intended.

This change is marked as breaking because we have overridden a
bunch of built-in methods (__eq__, __add__, etc.) to operate lazily.
The end result is that we have needed to rewrite test_linearize.py
quite significantly, and potentially other existing code that relied
on object identity checks might break.

However, I believe that the advantages in terms of usage ergonomics
and enabling gradual adoption of yakof features outweigh these
drawbacks. This pattern aligns us a bit more with TensorFlow v1.x.