feat(molveno): make the code testable
This diff edits the molveno/orig.py module to make the code
testable and ensure some basic level of testing.

Going forward, we'll need more of this and, specifically, we'll
need more testing. Yet, for now, it's good to start with some
assurance that, at least, the model is able to run.