feat(molveno): make the code testable (#57)
This diff edits the molveno/orig.py module to make the code testable and
ensure some basic level of testing.

Going forward, we'll need more of this and, specifically, we'll need
more testing. Yet, for now, it's good to start with some assurance that,
at least, the model is able to run.

Note that I sprinkled several `type: ignore` comments in there to avoid
to visually see the squiggles :grimacing:.

Also, we're going to see a coverage drop, but this is fine(TM). We're
adding new code that is known to be working as intended and that we wish
to change as little as possible, so it makes sense to be light on terms
of testing for now.