feat(dtyak): dt_model fork with relative imports (#58)
This commit adds a fork of the dt_model edited to use relative imports,
and reformatted using black.

The import commit is
https://github.com/fbk-most/dt-model/commit/efded93eee8956508faa265b3c5e071a48d11b23

For now, no tests and no guarantees this is working. I am trying to
follow a process where I can have clean diff with respect to the
original code in subsequent PRs.