feat(dtyak): use yakof.graph rather than sympy (#52)
This diff rewrites the `yakof/dtyak` package and the related example to
use `yakof.graph` as the graph building engine and `yakof.numpybackend`
as the numeric-evaluation backend. This diff gives us a feeling of what
needs to change in the original `dt_model` to use yakof code as the
underlying engine. Note that we don't support model substitution yet
and, in any case, I believe it would be more robust to steer away from
mutating the initial configuration.