feat(tests/molveno): add simple cross comparison test (#60)
This test A/B compares the two models using a fixed set of contextial
variables and scenarios. Directly comparing with more randomness is not
working as intended, but this may actually be down to differences in the
ensemble generation or randomness. In any case, a passing test case with
little randomness is a good starting point anyway.