fix: gtk-based matplotlib interactive plotting workaround
We have the issue that we cannot use tkinter with the python
version downloaded by `uv` as documented by https://github.com/astral-sh/python-build-standalone/issues/129,
therefore, interactive plotting is not working as intended.

As a workaround, document and add support for using the
`gtk` backend for `matplotlib`. This entails documenting
the system dependencies to install on a typical debian
system, and adding support for installing the gtk backend
with `uv sync`, should this be needed.

TBH, I am not satisfied by this solution, since it adds
friction that won't make researchers happy here. I'm
considering suggesting to either using a container or
recommending people to use LinuxBrew, which at least
should install Python binaries that WAI.

While there, ignore `/yakof.egg-info`, which I forgot to
add in my previous pull request.