fix: gtk-based matplotlib interactive plotting workaround (#65)
We have the issue that we cannot use tkinter with the python version
downloaded by `uv` as documented by
https://github.com/astral-sh/python-build-standalone/issues/129,
therefore, interactive plotting is not working as intended.

As a workaround, document and add support for using the `gtk` backend
for `matplotlib`. This entails documenting the system dependencies to
install on a typical debian system, and adding support for installing
the gtk backend with `uv sync --dev --group gtk`, should this be needed.

TBH, I am not satisfied by this solution, since it adds friction that
won't make researchers happy here. I'm considering suggesting to either
using a container or recommending people to use
[LinuxBrew](https://docs.brew.sh/Homebrew-on-Linux), which at least
should install Python binaries that work as intended. (I do not fully
understand *why* the binaries provided by `uv` do not include a dynamic
`_tkinter` dependency and I do not know why `matplotlib` requires to
check for `__file__`, which is optional, but anyway, here we are, with
either nonworking-by-default interactive plotting or uncertainty
regarding the Python version we're using.)

While there, ignore `/yakof.egg-info`, which I forgot to add in my
previous pull request.